+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Small refactoring
+- Return path of downloaded file
 
 ## [0.1.0] - 2022-07-10
 ### Added